+++ api/src/main/scala/pl/touk/nussknacker/engine/api/exception/NuExceptionInfo.scala
@@ -1,5 +1,6 @@
 package pl.touk.nussknacker.engine.api.exception
 
 import pl.touk.nussknacker.engine.api.Context
+import pl.touk.nussknacker.engine.api.component.NodeComponentInfo
 
-case class NuExceptionInfo[T <: Throwable](nodeId: Option[String], throwable: T, context: Context) extends Serializable
+case class NuExceptionInfo[T <: Throwable](nodeComponentInfo: Option[NodeComponentInfo], throwable: T, context: Context) extends Serializable