+++ engine/flink/executor/src/main/scala/pl/touk/nussknacker/engine/process/exception/FlinkExceptionHandler.scala
@@ -3,6 +3,7 @@ package pl.touk.nussknacker.engine.process.exception
 import com.typesafe.config.Config
 import net.ceedubs.ficus.Ficus.{booleanValueReader, optionValueReader, stringValueReader, toFicusConfig}
 import org.apache.flink.api.common.restartstrategy.RestartStrategies
+import pl.touk.nussknacker.engine.api.component.NodeComponentInfo
 import pl.touk.nussknacker.engine.api.exception.NuExceptionInfo
 import pl.touk.nussknacker.engine.api.process.ProcessObjectDependencies
 import pl.touk.nussknacker.engine.api.runtimecontext.EngineRuntimeContext
@@ -52,11 +53,11 @@ class FlinkExceptionHandler(metaData: MetaData,
     consumer.consume(extractOrThrow(exceptionInfo))
   }
 
-  override def handling[T](nodeId: Option[String], context: Context)(action: => T): Option[T] =
+  override def handling[T](nodeComponentInfo: Option[NodeComponentInfo], context: Context)(action: => T): Option[T] =
     try {
       Some(action)
     } catch {
-      case NonFatal(e) => handle(NuExceptionInfo(nodeId, e, context))
+      case NonFatal(e) => handle(NuExceptionInfo(nodeComponentInfo, e, context))
         None
     }