+++ engine/flink/executor/src/main/scala/pl/touk/nussknacker/engine/process/registrar/CollectingSinkFunction.scala
@@ -1,7 +1,8 @@
 package pl.touk.nussknacker.engine.process.registrar
 
 import org.apache.flink.streaming.api.functions.sink.{RichSinkFunction, SinkFunction}
-import pl.touk.nussknacker.engine.api.{InterpretationResult, ValueWithContext}
+import pl.touk.nussknacker.engine.api.component.{ComponentType, NodeComponentInfo}
+import pl.touk.nussknacker.engine.api.ValueWithContext
 import pl.touk.nussknacker.engine.process.ExceptionHandlerFunction
 import pl.touk.nussknacker.engine.process.compiler.FlinkProcessCompilerData
 import pl.touk.nussknacker.engine.testmode.SinkInvocationCollector
@@ -11,7 +12,7 @@ private[registrar] class CollectingSinkFunction[T](val compiledProcessWithDepsPr
   extends RichSinkFunction[ValueWithContext[T]] with ExceptionHandlerFunction {
 
   override def invoke(value: ValueWithContext[T], context: SinkFunction.Context): Unit = {
-    exceptionHandler.handling(Some(sinkId), value.context) {
+    exceptionHandler.handling(Some(NodeComponentInfo(sinkId, "collectingSinkFunction", ComponentType.Sink)), value.context) {
       collectingSink.collect(value.context, value.value)
     }
   }