+++ engine/lite/api/src/main/scala/pl/touk/nussknacker/engine/lite/api/utils/errors.scala
@@ -1,6 +1,7 @@
 package pl.touk.nussknacker.engine.lite.api.utils
 
 import pl.touk.nussknacker.engine.api.Context
+import pl.touk.nussknacker.engine.api.component.{ComponentInfo, NodeComponentInfo}
 import pl.touk.nussknacker.engine.api.exception.NuExceptionInfo
 import pl.touk.nussknacker.engine.lite.api.commonTypes.ErrorType
 import pl.touk.nussknacker.engine.lite.api.customComponentTypes.CustomComponentContext
@@ -10,8 +11,8 @@ import scala.util.Try
 
 object errors {
 
-  def withErrors[F[_], T](customComponentContext: CustomComponentContext[F], ctx: Context)(action: => T): Either[ErrorType, T] = {
-    Try(action).toEither.left.map(NuExceptionInfo(Some(customComponentContext.nodeId), _, ctx))
+  def withErrors[F[_], T](customComponentContext: CustomComponentContext[F], componentInfo: Option[ComponentInfo], ctx: Context)(action: => T): Either[ErrorType, T] = {
+    Try(action).toEither.left.map(NuExceptionInfo(Some(NodeComponentInfo(customComponentContext.nodeId, componentInfo)), _, ctx))
   }
 
 }