+++ engine/lite/request-response/runtime/src/test/scala/pl/touk/nussknacker/engine/requestresponse/RequestResponseConfigCreator.scala
@@ -13,12 +13,12 @@ import pl.touk.nussknacker.engine.api.signal.ProcessSignalSender
 import pl.touk.nussknacker.engine.api.test.InvocationCollectors.ServiceInvocationCollector
 import pl.touk.nussknacker.engine.api.typed.typing.Typed
 import pl.touk.nussknacker.engine.lite.api.commonTypes._
-import pl.touk.nussknacker.engine.lite.api.customComponentTypes.{LiteCustomComponent, CustomComponentContext}
+import pl.touk.nussknacker.engine.lite.api.customComponentTypes.{CustomComponentContext, LiteCustomComponent}
 import pl.touk.nussknacker.engine.lite.api.utils.sinks.LazyParamSink
 import pl.touk.nussknacker.engine.lite.api.utils.transformers.SingleElementComponent
 import pl.touk.nussknacker.engine.requestresponse.api.RequestResponseSinkFactory
 import pl.touk.nussknacker.engine.requestresponse.utils.customtransformers.Sorter
-import pl.touk.nussknacker.engine.requestresponse.utils.{JsonSchemaRequestResponseSourceFactory, JsonRequestResponseSourceFactory}
+import pl.touk.nussknacker.engine.requestresponse.utils.{JsonRequestResponseSourceFactory, JsonSchemaRequestResponseSourceFactory}
 import pl.touk.nussknacker.engine.util.LoggingListener
 import pl.touk.nussknacker.engine.util.service.{EnricherContextTransformation, TimeMeasuringService}