+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/Interpreter.scala
@@ -12,6 +12,7 @@ import pl.touk.nussknacker.engine.api.process.RunMode
 import pl.touk.nussknacker.engine.compiledgraph.node.{Sink, Source, _}
 import pl.touk.nussknacker.engine.compiledgraph.service._
 import pl.touk.nussknacker.engine.compiledgraph.variable._
+import pl.touk.nussknacker.engine.component.NodeComponentInfoExtractor
 import pl.touk.nussknacker.engine.expression.ExpressionEvaluator
 import pl.touk.nussknacker.engine.util.SynchronousExecutionContext
 
@@ -44,7 +45,9 @@ private class InterpreterInternal[F[_]](listeners: Seq[ProcessListener],
 
   private implicit def nodeToId(implicit node: Node): NodeId = NodeId(node.id)
 
-  private def handleError(node: Node, ctx: Context): Throwable => NuExceptionInfo[_ <: Throwable] = NuExceptionInfo(Some(node.id), _, ctx)
+  private def handleError(node: Node, ctx: Context): Throwable => NuExceptionInfo[_ <: Throwable] = {
+    NuExceptionInfo(Some(NodeComponentInfoExtractor.fromNode(node)), _, ctx)
+  }
 
   private def interpretNode(node: Node, ctx: Context): F[List[Result[InterpretationResult]]] = {
     implicit val nodeImplicit: Node = node
@@ -55,7 +58,7 @@ private class InterpreterInternal[F[_]](listeners: Seq[ProcessListener],
       case _ => listeners.foreach(_.nodeEntered(node.id, ctx, metaData))
     }
     node match {
-      case Source(_, next) =>
+      case Source(_, _, next) =>
         interpretNext(next, ctx)
       case VariableBuilder(_, varName, Right(fields), next) =>
         val variable = createOrUpdateVariable(ctx, varName, fields)