+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/compiledgraph/node.scala
@@ -13,7 +13,7 @@ object node {
     def id: String
   }
 
-  case class Source(id: String, next: Next) extends Node
+  case class Source(id: String, ref: Option[String], next: Next) extends Node
 
   case class Sink(id: String, ref: String, isDisabled: Boolean) extends Node