+++ utils/util/src/main/scala/pl/touk/nussknacker/engine/util/exception/WithExceptionExtractor.scala
@@ -1,6 +1,6 @@
 package pl.touk.nussknacker.engine.util.exception
 
-import pl.touk.nussknacker.engine.api.exception.{NuExceptionInfo, NonTransientException}
+import pl.touk.nussknacker.engine.api.exception.{NonTransientException, NuExceptionInfo}
 import pl.touk.nussknacker.engine.api.util.ReflectUtils
 import pl.touk.nussknacker.engine.util.exception.WithExceptionExtractor.{DefaultNonTransientExceptionExtractor, DefaultTransientExceptionExtractor}
 import pl.touk.nussknacker.engine.util.logging.LazyLoggingWithTraces
@@ -29,12 +29,12 @@ trait WithExceptionExtractor extends LazyLoggingWithTraces {
       case transientExceptionExtractor(_) =>
         throw exceptionInfo.throwable
       case nonTransientExceptionExtractor(nonTransient) =>
-        NuExceptionInfo(exceptionInfo.nodeId, nonTransient, exceptionInfo.context)
+        NuExceptionInfo(exceptionInfo.nodeComponentInfo, nonTransient, exceptionInfo.context)
       case other =>
         val exceptionDetails = s"${ReflectUtils.simpleNameWithoutSuffix(other.getClass)}:${other.getMessage}"
         val nonTransient = NonTransientException(input = exceptionDetails, message = "Unknown exception", cause = other)
         infoWithDebugStack(s"Unknown exception $exceptionDetails for ${exceptionInfo.context.id}", other)
-        NuExceptionInfo(exceptionInfo.nodeId, nonTransient, exceptionInfo.context)
+        NuExceptionInfo(exceptionInfo.nodeComponentInfo, nonTransient, exceptionInfo.context)
     }
   }